services/python-collector/requirements.txt
# ARBITRAGEXPLUS2025 - Python Collector Dependencies
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
pydantic==2.5.3
redis==5.0.1
psycopg2-binary==2.9.9
schedule==1.2.0
pytz==2023.3
